backend/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
drf-yasg==1.21.6
python-dotenv
coreapi

